moved user handling from utilityFunctions to userHandler
