+++ src/components/OngoingEventScreens/SingleTargetScreen.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { View, Linking } from 'react-native'
 import { Text, Button } from 'react-native-elements'
 import styles, { paddingSides } from '../../styles/global'
+import { connect } from 'react-redux'
 import { KYPPI_URL } from 'react-native-dotenv'
 import colors from '../../styles/colors'
 import { selectTarget, setSelectedTargets } from '../../reducers/targetReducer'
-import { connect } from 'react-redux'
 
 const style = {
   container: {
